extract common merge function
