+++ pkg/querier/tripperware/queryrange/query_range.go
@@ -21,7 +21,6 @@ import (
 	"github.com/prometheus/prometheus/model/timestamp"
 	"github.com/weaveworks/common/httpgrpc"
 
-	"github.com/cortexproject/cortex/pkg/cortexpb"
 	"github.com/cortexproject/cortex/pkg/querier/tripperware"
 	"github.com/cortexproject/cortex/pkg/util"
 	"github.com/cortexproject/cortex/pkg/util/spanlogger"
@@ -345,32 +344,12 @@ func statsMerge(shouldSumStats bool, resps []*PrometheusResponse) *tripperware.P
 }
 
 func matrixMerge(resps []*PrometheusResponse) []tripperware.SampleStream {
-	output := map[string]*tripperware.SampleStream{}
+	output := make(map[string]tripperware.SampleStream)
 	for _, resp := range resps {
-		for _, stream := range resp.Data.Result {
-			metric := cortexpb.FromLabelAdaptersToLabels(stream.Labels).String()
-			existing, ok := output[metric]
-			if !ok {
-				existing = &tripperware.SampleStream{
-					Labels: stream.Labels,
-				}
-			}
-			// We need to make sure we don't repeat samples. This causes some visualisations to be broken in Grafana.
-			// The prometheus API is inclusive of start and end timestamps.
-			if len(existing.Samples) > 0 && len(stream.Samples) > 0 {
-				existingEndTs := existing.Samples[len(existing.Samples)-1].TimestampMs
-				if existingEndTs == stream.Samples[0].TimestampMs {
-					// Typically this the cases where only 1 sample point overlap,
-					// so optimize with simple code.
-					stream.Samples = stream.Samples[1:]
-				} else if existingEndTs > stream.Samples[0].TimestampMs {
-					// Overlap might be big, use heavier algorithm to remove overlap.
-					stream.Samples = tripperware.SliceSamples(stream.Samples, existingEndTs)
-				} // else there is no overlap, yay!
-			}
-			existing.Samples = append(existing.Samples, stream.Samples...)
-			output[metric] = existing
+		if resp == nil {
+			continue
 		}
+		tripperware.MergeSampleStreams(output, resp.Data.GetResult())
 	}
 
 	keys := make([]string, 0, len(output))
@@ -381,7 +360,7 @@ func matrixMerge(resps []*PrometheusResponse) []tripperware.SampleStream {
 
 	result := make([]tripperware.SampleStream, 0, len(output))
 	for _, key := range keys {
-		result = append(result, *output[key])
+		result = append(result, output[key])
 	}
 
 	return result
