+++ pkg/querier/tripperware/util.go
@@ -3,11 +3,9 @@ package tripperware
 import (
 	"context"
 	"net/http"
-	"sort"
 
 	"github.com/weaveworks/common/httpgrpc"
 
-	"github.com/cortexproject/cortex/pkg/cortexpb"
 	"github.com/cortexproject/cortex/pkg/tenant"
 	"github.com/cortexproject/cortex/pkg/util/validation"
 )
@@ -71,25 +69,4 @@ func DoRequests(ctx context.Context, downstream Handler, reqs []Request, limits 
 	}
 
 	return resps, firstErr
-}
-
-// SliceSamples assumes given samples are sorted by timestamp in ascending order and
-// return a sub slice whose first element's is the smallest timestamp that is strictly
-// bigger than the given minTs. Empty slice is returned if minTs is bigger than all the
-// timestamps in samples.
-// If the given samples slice is not sorted then unexpected samples would be returned.
-func SliceSamples(samples []cortexpb.Sample, minTs int64) []cortexpb.Sample {
-	if len(samples) == 0 || minTs < samples[0].TimestampMs {
-		return samples
-	}
-
-	if len(samples) > 0 && minTs > samples[len(samples)-1].TimestampMs {
-		return samples[len(samples):]
-	}
-
-	searchResult := sort.Search(len(samples), func(i int) bool {
-		return samples[i].TimestampMs > minTs
-	})
-
-	return samples[searchResult:]
 }
